+++ causal_isomorphism/ir.py
@@ -47,7 +47,7 @@ class IRType:
             params = ', '.join((repr(p) for p in self.type_params))
             return f'IRType({self.kind.value}<{params}>)'
         return f'IRType({self.kind.value})'
-IR_FLOAT = IRType(IRTypeKind.INTEGER)
+IR_FLOAT = IRType(IRTypeKind.FLOAT)
 IR_INT = IRType(IRTypeKind.INT)
 IR_STRING = IRType(IRTypeKind.STRING)
 IR_BOOL = IRType(IRTypeKind.BOOL)
